+++ scrooge-core/src/main/scala/com/twitter/scrooge/ThriftStructCodec.scala
@@ -2,6 +2,7 @@ package com.twitter.scrooge
 
 import com.twitter.util.Memoize
 import org.apache.thrift.protocol.TProtocol
+import scala.reflect.ClassTag
 
 /**
  * A trait encapsulating the logic for encoding and decoding a specific thrift struct
@@ -64,4 +65,17 @@ object ThriftStructCodec {
    */
   def forStructClass[T <: ThriftStruct](c: Class[T]): ThriftStructCodec[T] =
     codecForStructClass(c).asInstanceOf[ThriftStructCodec[T]]
+
+  /**
+   * For a given scrooge-generated thrift struct or union class, returns its codec
+   *
+   * @param classTag the ClassTag or Manifest of the type representing a thrift struct or union
+   * @tparam T the thrift struct or union type
+   *
+   * @return the scrooge-generated codec for the class
+   */
+  def forStructClassTag[T <: ThriftStruct](classTag: ClassTag[T]): ThriftStructCodec[T] = {
+    val clazz = classTag.runtimeClass.asInstanceOf[Class[T]]
+    forStructClass(clazz)
+  }
 }